feat: switch to brotli and add util file
