+++ Cargo.toml
@@ -1,2 +1,11 @@
 [workspace]
 members = ["test_crates/*"]
+
+[profile.release]
+codegen-units = 1
+opt-level = "z"
+lto = true
+debug = false
+panic = "abort"
+overflow-checks = true
+
