+++ package.json
@@ -36,6 +36,7 @@
   },
   "homepage": "https://github.com/TENK-DAO/nft-cli",
   "devDependencies": {
+    "@types/brotli": "^1.3.1",
     "@types/jest": "^27.0.3",
     "@types/node": "^16.11.13",
     "@types/yargs": "^17.0.7",
@@ -52,14 +53,12 @@
     "jest": "^27.4.5",
     "prettier": "^2.5.1",
     "ts-jest": "^27.1.1",
+    "ts-node": "^10.7.0",
     "typescript": "^4.5.4"
   },
   "dependencies": {
-    "@ipld/car": "^4.0.0",
-    "@pinata/sdk": "^1.1.25",
     "bn.js": "^5.2.0",
-    "ipfs-car": "^0.6.2",
-    "nft.storage": "^6.0.2",
+    "brotli": "^1.3.2",
     "witme": "^0.2.5",
     "yargs": "^17.3.0"
   },
